requirements.txt
click==7.1.2
pyelftools==0.27
dotmap~=1.3.23